splits test_web
